demo/requirements-demo.txt
PyYAML>=6.0
attr>=0.3.2
gym==0.26.2
imageio>=2.37.2
jsonlines>=4.0.0
matplotlib>=3.8
nbformat>=5.10
networkx>=3.3
nltk>=3.9
notebook>=7.0
numpy>=1.26
openai>=1.0
pandas>=2.2
python-dotenv>=1.0
